// src/modules/centers/enums/center-activity-type.enum.ts
export enum CenterActivityType {
  CENTER_CREATED = 'CENTER_CREATED',
  CENTER_UPDATED = 'CENTER_UPDATED',
  CENTER_DELETED = 'CENTER_DELETED',
  CENTER_RESTORED = 'CENTER_RESTORED',
  CENTER_ACCESS_GRANTED = 'CENTER_ACCESS_GRANTED',
  CENTER_ACCESS_REVOKED = 'CENTER_ACCESS_REVOKED',
  BRANCH_CREATED = 'BRANCH_CREATED',
  BRANCH_UPDATED = 'BRANCH_UPDATED',
  BRANCH_DELETED = 'BRANCH_DELETED',
}