// src/modules/user/enums/user-activity-type.enum.ts
export enum UserActivityType {
  USER_CREATED = 'USER_CREATED',
  USER_UPDATED = 'USER_UPDATED',
  USER_DELETED = 'USER_DELETED',
  USER_RESTORED = 'USER_RESTORED',
  USER_ACTIVATED = 'USER_ACTIVATED',
  USER_DEACTIVATED = 'USER_DEACTIVATED',
  USER_ACCESS_GRANTED = 'USER_ACCESS_GRANTED',
  USER_ACCESS_REVOKED = 'USER_ACCESS_REVOKED',
  USER_IMPORTED = 'USER_IMPORTED',
}