// src/modules/access-control/enums/role-activity-type.enum.ts
export enum RoleActivityType {
  ROLE_CREATED = 'ROLE_CREATED',
  ROLE_UPDATED = 'ROLE_UPDATED',
  ROLE_DELETED = 'ROLE_DELETED',
  ROLE_RESTORED = 'ROLE_RESTORED',
  ROLE_ASSIGNED = 'ROLE_ASSIGNED',
  ROLE_REMOVED = 'ROLE_REMOVED',
  PERMISSION_ADDED_TO_ROLE = 'PERMISSION_ADDED_TO_ROLE',
  PERMISSION_REMOVED_FROM_ROLE = 'PERMISSION_REMOVED_FROM_ROLE',
  PERMISSION_SCOPE_CHANGED = 'PERMISSION_SCOPE_CHANGED',
}